+++ ciri/__init__.py
@@ -2,4 +2,4 @@ from ciri.core import Schema, SchemaOptions, PolySchema
 from ciri.exception import ValidationError
 
 
-__version__ = '0.2.1'
+__version__ = '0.6.0'